requirements.txt
PyMySQL==1.1.1
beautifulsoup4==4.13.3
pandas==2.2.3
matplotlib==3.10.1
SQLAlchemy==2.0.38
python-dotenv==1.0.1
cryptography==44.0.2
ipykernel==6.29.5
langchain==0.3.20
tiktoken==0.9.0
weaviate-client==4.11.1
sentence-transformers==3.4.1
pyvi==0.1.1
tqdm==4.67.1
langchain-experimental==0.3.4
langchain-community==0.3.19
langchain-huggingface==0.1.2
# google-generativeai==0.8.4
langchain-google-genai==2.0.10